requirements.txt
Django==3.0.1
djangorestframework==3.11.0
psycopg2-binary==2.8.4
python-dotenv==0.10.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
